Added flexibility to time parameter definition; changed time-dependent solver imports
